+++ voicenavigator-20180612/ChangeLog.md
@@ -1,3 +1,16 @@
+2024-02-04 Version: 1.0.3
+- Update API AssociateChatbotInstanceadd NluServiceType param.
+update ChatbotInstanceId param.
+- Update API AuditTTSVoiceadd AccessKey param.
+add AppKey param.
+add Engine param.
+add PitchRate param.
+add SecretKey param.
+- Update API DescribeTTSConfigupdate response param.
+- Update API ModifyTTSConfigadd Engine param.
+add EngineXunfei param.
+
+
 2023-10-16 Version: 1.0.2
 - Generated python 2018-06-12 for VoiceNavigator.